Setup authentication

+++ laravel_backend/database/migrations/2022_06_17_033647_add_role_to_users_table.php
@@ -14,7 +14,7 @@ class AddRoleToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('role_code');
+            $table->string('role_code')->after('email_verified_at');
             $table->foreign('role_code')->references('code')->on('roles');
         });
     }
@@ -26,8 +26,11 @@ class AddRoleToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role_code');
-        });
+        if (Schema::hasColumn('users', 'role_code')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropForeign(['role_code']);
+                $table->dropColumn('role_code');
+            });
+        }
     }
 }
